根据菜单id查找 页面回显代码开发

+++ sky-server/src/main/java/com/sky/mapper/SetmealMapper.java
@@ -53,8 +53,8 @@ public interface SetmealMapper {
             "#{id}" +
             "</foreach>" +
             "</script>"})
-    void deleteById(List<Integer> ids);
+    void deleteById(List<Long> ids);
 
     @Select("select * from setmeal where id = #{id}")
-    Setmeal getById(Integer id);
+    Setmeal getById(Long id);
 }
